LF-5041 undo backed isLibrary commenting -- it is causing no harm

+++ packages/api/src/controllers/productController.js
@@ -123,9 +123,8 @@ const productController = {
         }
 
         // LF-4963 - confirm property that will distinguish custom from library products
-        // const isLibraryProduct = product.product_translation_key;
-        // if (!isLibraryProduct && isUnusedByTasks) {
-        if (isUnusedByTasks) {
+        const isLibraryProduct = product.product_translation_key;
+        if (!isLibraryProduct && isUnusedByTasks) {
           await baseController.delete(ProductModel, product_id, req, { trx });
         }
